fix: reuse visualizationGroup component
