+++ src/components/analyticVisualization/VisualizationGroup/VisualizationGroup.js
@@ -1,11 +1,12 @@
-import React, { useState } from 'react'
+import React, {useEffect, useState} from 'react'
 import PropTypes from '@dhis2/prop-types'
 import i18n from '@dhis2/d2-i18n'
 import find from 'lodash/find'
 import isEmpty from 'lodash/isEmpty'
-import { CheckboxField } from '../../field'
+import {CheckboxField, TextField} from '../../field'
 import { GroupType } from './GroupType'
 import { useSystemId } from '../../../utils/useSystemId'
+import styles from "./styles/GroupField.module.css";
 
 const DEFAULT = 'default'
 
@@ -26,10 +27,37 @@ const findDefaultGroup = (type, groupList, settings) => {
     return defaultGroup
 }
 
-export const VisualizationGroup = ({ settings, onChange, groupList, type }) => {
+const isDefaultGroup = group => group.name === 'default'
+
+const findGroup = (type, group, settings) => {
+    const { groupList, groupId } = group
+
+    switch (type) {
+        case 'program':
+        case 'dataset':
+            return settings.group
+        default:
+            return find(groupList, g => g.id === groupId)
+    }
+}
+
+export const VisualizationGroup = ({ settings, onChange, groupList, type, groupId, disabled }) => {
     const { data: id } = useSystemId()
     const [group, setGroup] = useState(true)
     const [groupType, setGroupType] = useState(true)
+    const [title, setTitle] = useState('')
+
+    useEffect(() => {
+        const groupFound = findGroup(type, { groupList, groupId }, settings)
+        if (disabled) {
+            if (!isDefaultGroup(groupFound)) {
+                setGroup(!!groupFound)
+                setTitle(groupFound.name)
+            } else {
+                setGroup(false)
+            }
+        }
+    }, [groupList])
 
     const handleChange = e => {
         setGroup(e.checked)
@@ -53,15 +81,22 @@ export const VisualizationGroup = ({ settings, onChange, groupList, type }) => {
                 label={i18n.t('Use a group visualization')}
                 checked={group}
                 onChange={handleChange}
+                disabled={disabled}
             />
 
-            {groupType && (
-                <GroupType
-                    onChange={onChange}
-                    settings={settings}
-                    groupList={groupList}
-                    type={type}
-                />
+            {disabled ?
+                group && (
+                    <div className={styles.field}>
+                        <TextField dense value={title} disabled />
+                    </div>
+                )
+                : groupType && (
+                    <GroupType
+                        onChange={onChange}
+                        settings={settings}
+                        groupList={groupList}
+                        type={type}
+                    />
             )}
         </>
     )
